requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
discord.py==2.1.0
dropbox==11.26.0
emoji==1.4.1
googlemaps==4.5.1
lxml==4.6.3
openai==0.27.0
pandas-datareader==0.10.0
pykakasi==2.2.1
numpy==1.21.1
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
speedtest-cli==2.1.2
wikipedia==1.4.0
